ai_service/requirements.txt
# Core ML Libraries
tensorflow>=2.18.0
scikit-learn>=1.2.2
numpy>=1.23.5
pandas>=2.0.1
joblib>=1.2.0

# Data Processing
python-dateutil>=2.8.2
pytz>=2023.3

# API and Web
fastapi>=0.95.2
uvicorn>=0.22.0
python-dotenv>=1.0.0
requests>=2.30.0
aiohttp>=3.8.4

# Settings and Validation
pydantic>=2.10.4
pydantic-settings>=2.7.1

# Redis for Caching
redis>=4.5.4
hiredis>=2.2.3

# Testing and Development
pytest>=7.3.1
pytest-asyncio>=0.21.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0

# Monitoring and Logging
prometheus-client>=0.16.0
python-json-logger>=2.0.7

# Optional: GPU Support
# tensorflow-gpu>=2.18.0  # Uncomment if using GPU

# Performance Optimization
numba>=0.57.0
bottleneck>=1.3.7

# Technical Analysis
ta-lib>=0.4.25  # Note: Requires system-level installation of TA-Lib
pandas-ta>=0.3.14b